LayoutTests/http/tests/security/contentSecurityPolicy/1.1/frame-ancestors/frame-ancestors-nested-same-in-same-none-block-expected.txt

CONSOLE MESSAGE: Refused to display 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/frame-ancestors.pl?policy='none'' in a frame because an ancestor violates the following Content Security Policy directive: "frame-ancestors 'none'".

A 'frame-ancestors' CSP directive with a value 'none' should block rendering in nested frames. This test PASSED if you do not see any content in the inner-most nested iframe below. Otherwise, it FAILED.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
Testing a same-origin child with a policy of "'none'" nested in a same-origin parent.



--------
Frame: '<!--framePath //<!--frame0-->/<!--frame0-->-->'
--------
